Generated classes based on AndroidAnnotations on Gradle compile
This makes use of the "full" JAR of AndroidAnnotations which is currently hard-coded
to 2.2.  The AndroidAnnotations API should be declared as a dependency in the
build.gradle file that uses it, as I haven't figured out yet how to get it to work
without defining it.  The full JAR should also be included in the lib/ folder.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>