Enable annotation processing with AndroidAnnotations with IDEA projects
This adds additional entries to IDEA' IPR XML, assuming that the "idea" plugin
is applied to a build script.

Currently this expects the "full" AndroidAnnotations JAR in lib and is of
version 2.2.  The module name is also hard-coded to "HelloWorld" at the moment,
for the integration test to pass.   These should be configurable once we
add a convention to this plugin.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>