Move AndroidAnnotations dependencies configuration to the plugin itself
Before, you need to define the AndroidAnnotations in the dependencies
section of the build script, as the plugin does not do this for you.  I
thought that was not possible as running the integration test in
1.0-milestone-6 took a long time, so maybe it didn't work (and I wanted
to get 0.1.0 out).

For 1.0-milestone-7 the integration test was completed much faster, which
made me realize that it is indeed possible to have the dependencies
section in the plugin.  And as a pre-requisite, the repositories section
was also moved to the plugin.

This means that the only things that need to be added to a build script
are the buildscript section, and `apply plugin: androidannotations`.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>