Define AndroidAnnotations "full" JAR in dependencies, so we no longer have to add it to the source manually
AndroidAnnotations comes in 2 JARs: API, which contains the annotationsi only
and is meant to be package with Android projects, and "full", which contains
the annotation processor and other support classes.  Full is a superset of
API, which means that they couldn't be loaded at the same time as that will
cause duplicate class file errors.

We worked around this before by simply defining only the API JAR in the
dependencies section.  The full JAR is added manually.  While this works it
takes longer to setup a AndroidAnnotations-based Gradle project, and makes
upgrading AndroidAnnotations versions much harder.

The key problem to solve is ensuring that both API and full JARs are not in
the same classpath.  We do this by creating a custom configuration
"androidannotations" that will contain the full JAR.  We thus make sure that
our Java and IDEA plugins scan the full JAR from the androidannotations
classpath in order for them to process the annotations correctly.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>