Use new package of OperatingSystem class in 1.0-milestone-8
The OperationgSystem class was moved from the os package to internal.nativeplatform
in 1.0-milestone-8, which caused the integration tests to fail in the new Gradle
version.  We now use the new package which unfortunately means that we need to have
at least 1.0-milestone-8 to develop on this.

Ideally we should remove our dependency to the OperatingSystem class, which I should
be doing sometime soon.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>